+++ src/controllers/MoovieController.js
@@ -20,6 +20,7 @@ const MooviesController = {
     async find(req, res) {
         const { id } = req.params;
         try {
+            
             const moovies = await db.query(`
             SELECT 
               m.*,
@@ -45,6 +46,14 @@ const MooviesController = {
         const { title, description, category_id, realease_date } = req.body;
 
         //é necessario realizar um validaçao pelo id de categoria
+        const ValCategory = await db.query(
+            `SELECT * FROM category WHERE id = $1`,
+            [category_id]
+        );
+        if (ValCategory.rows.length === 0) {
+            return res.status(400).json({ error: "Categoria nao encontrada" });
+        }
+
         try {
             const newMoovie = await db.query(
                 `INSERT INTO  moovie (title, description, category_id, realease_date)
@@ -61,19 +70,61 @@ const MooviesController = {
     async delete(req, res) {
         const { id } = req.params;
         try {
+            const ValId = await db.query(
+                `SELECT * FROM moovie WHERE id = $1`,
+                [id]
+            );
+            if (ValId.rows.length === 0) {
+                return res.status(404).json({ error: "Filme nao encontrado" });
+            }
+
             const result = await db.query(
                 "DELETE FROM moovie WHERE id = $1 RETURNING *",
                 [id]
             );
             if (result.rowCount > 0) {
                 res.status(204).json({});
-            }else{
+            } else {
                 res.status(304).json({});
             }
         } catch (error) {
             res.status(500).json({ error: error.message });
         }
     },
+
+    async update(req, res) {
+        const { id } = req.params;
+        const { title, description, category_id, realease_date } = req.body;
+
+        try {
+            const ValCategory = await db.query(
+                `SELECT * FROM category WHERE id = $1`,
+                [category_id]
+            );
+            if (ValCategory.rows.length === 0) {
+                return res.status(404).json({ error: "Categoria nao encontrada" });
+            }
+
+            const updateMoovie = await db.query(
+                `UPDATE moovie 
+            SET title = $1, description = $2, category_id = $3, realease_date = $4 
+            WHERE id = $5
+            RETURNING *;`,
+                [title, description, category_id, realease_date, id]
+            );
+
+            if (updateMoovie.rowCount > 0) {
+                res.status(200).json(updateMoovie.rows[0]);
+            } else {
+                res.status(304).json({ error: "atualização nao encontrada" });
+            }
+
+
+        } catch (error) {
+            res.status(500).json({ error: error.message });
+        }
+    }
 }
 
 module.exports = MooviesController;
+
